requirements.txt
cohere
python-dotenv
alpaca-py
pyautogen
requests